requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.6
numpy==1.24.4
opencv-python==4.9.0.80
packaging==24.0
Pillow==9.5.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.31.0
typing_extensions==4.11.0
urllib3==2.2.1
